Adds and tests ability for user to select a contractor and update user_choice
<42710030+tnodland@users.noreply.github.com>
