+++ fix_up/urls.py
@@ -8,6 +8,7 @@ from .views import ListProjectsByContractor
 from .views import ListProjectsByUser
 from .views import ListProjectBatchView
 from .views import SwipeUpdateContractorChoiceView
+from .views import UpdateUserChoiceView
 
 urlpatterns = [
     path('contractors/', CreateContractorView.as_view()),
@@ -18,5 +19,6 @@ urlpatterns = [
     path('contractors/<int:contractor_id>/projects', ListProjectsByContractor.as_view()),
     path('users/<int:user_id>/projects', ListProjectsByUser.as_view()),
     path('projects', ListProjectBatchView.as_view()),
-    path('contractors/<int:contractor_id>/projects/<int:project_id>', SwipeUpdateContractorChoiceView.as_view())
+    path('contractors/<int:contractor_id>/projects/<int:project_id>', SwipeUpdateContractorChoiceView.as_view()),
+    path('projects/<int:project_id>/contractors/<int:contractor_id>', UpdateUserChoiceView.as_view())
 ]
